[not verified] introduce ai assistant bar cmp
